+++ py/src/__main__.py
@@ -1,5 +1,4 @@
-import app
-
+from src.desktop import app
 
 app = app.App()
 app.mainloop()
