+++ py/src/desktop/app.py
@@ -1,7 +1,7 @@
 import tkinter
 
-from window import Window
-from login import LoginWindow
+from .window import Window
+from .login import LoginWindow
 
 from typing import Union
 from typing import Type
